fix onTapGesture bounds
